Use Test::More functions instead of hand-crafted ones
In order to eliminate the special workarounds for "no-dot-in-@INC" in
t/TEST, we need to have fixes applied to the upstream distributions.

This commit begins the work of fixing cpan/libnet.  Four test files in
this distribution use hand-crafted testing subroutines found in
t/libnet_t.pl.  The four test files "require t/libnet_t.pl" -- which is
what runs afoul of no-dot-in-@INC.

However, those testing subroutines appear to date from the very early
days of Perl 5 when Test::More had not yet been written.  The Test::More
functions are drop-in replacements for the t/libnet_t.pl ones.