Use Test::More::skip_all() where prereqs not met
Smoke-testing note:  This branch is forked from a branch which has
already eliminated the t/TEST workaround with respect to
cpan/Test-Simple.  However, there is one test file in that distro which
is still failing on Windows (though not Linux).  That is under
investigation.  Until that is resolved, this branch will show a test
failure on Windows for cpan/Test-Simple -- but what we're interested in
is results for cpan/libnet.