Re-order $Config{libpth} in DynaLoader in 64bit env
In a 64bit environment, force 64bit locations *before* 32bit locations.
This will move /usr/lib64 in front of /usr/lib and /lib64 in front of
/lib.  This is important if a module uses $Config{libpth} to find a
library to load, like FFI::CheckLib does. If a library is installed in
both locations, it is likely to pick the wrong one if the order is not
showing the correct architecture first
As this clutch only looks at trailing 64, it will have no impact on
locations differing completely, as /usr/lib vs /usr/lib/i386-linux-gnu/
like on debian.

Also just include existing PATH elements and filter duplicates.