Update hints/os390.sh
This is an extensive reworking of this hints file that was a mismash of
tribal knowledge from 20 years ago.

Some things were removed becaucse Configure is now smarter; for example
it now tests gcvt for the failures that this assumed, and doesn't use
it, without the hints having to know about it.

I went through the IBM documentation for each option, and found lots of
redundancies.  For example, we were specifying both C89 and C99
behavior.  Various things started working properly without having to
further specify things, when just C99 is said.  Similarly _ALL_SOURCE
encompasses a bunch of other options we were specifying individually.

This OS doesn't support POSIX 2008, but supposedly does a POSIX 2004
(which I had never heard of), but using that broke things.

I commented the things I found, so the reasons behind them don't have to
be guessed at.  There are a few things I didn't find the reasons behind.