lib/vars.t: Adapt to using harness  TODO
The file as it currently exists is very difficult to work with.  The
test numbers are hard-coded and are not amenable to having descriptions.
So it's difficult to figure out where you are in the file if you are
debugging an error.

There is one place in the original where I haven't yet figured out what
the point of the test is.  So we're running one test short of the number
in the plan.