Revamp regcurly(), regpiece() use of it
This commit copies portions of new_regcurly(), which has been around
since 5.28, into plain regcurly(), as a baby step in preparation for
converting entirely to the new one.  These functions are used for
parsing {m,n} quantifiers.  Future commits will add capabilities not
available using the old version.

The commit adds an optional parameter, to return to the caller
information it gleans during parsing.

regpiece() is changed by this commit to use this information, instead of
itself reparsing the input.  Part of the reason for this commit is that
changes are planned soon to what is legal syntax.  With this commit in
place, those changes only have to be done once.

This commit also extracts into a function the calculation of the
quantifier bounds.  This allows the logic for that to be done in one
place instead of two.