Revert "Don't test for 16-bit inputs in inRANGE()"
This reverts commit 4c2aa7c802893d0276551ade1b9d5dcd1226afc4.

That commit removed checking for 16 bit inputs because an assert in a
macro was making compilations fail in compilers that have a small macro
buffer.  It would have been better to treat such compilers specially,
because it turns out that systems which use UTF-16, like Mingw and Win32
do need that.

A mistake made in e7ae132ec78 caused all Win32 compiler versions to be
so-treated, and so they continued to compile.  But some change in 5.32
caused Mingw to run afoul of this problem; but since it's an assert, it
only affected debugging builds, which were tested much less often.

The reverted commit mentioned HP compilers as having problems.  The only
such box we currently have available for testing is HP-31, and the
reversion works fine on it.  This may break older versions, and if that
happens, the solution is to add them to the small buffer class.