Prevent rare locale race
In reading code, I realized that it is possible for there to be a race
here if two threads operating in a Posix 2008 locale system convert to
the old-style locale system.  That is unlikely, but it is easy to
prevent any race through a mutex.  They still would have problems if
they do a tug-of-war on what the global locale should be.